Codeql?

+++ rfc1035.h
@@ -283,7 +283,7 @@ namespace RFC1035
         r[p++] = l.size();
         if (p == 0)
           throw std::runtime_error("name too long for small_vector");
-        for (uint8_t i = 0; i < l.size(); i++)
+        for (size_t i = 0; i < l.size(); i++)
         {
           r[p++] = l[i];
           if (p == 0)
